+++ R/5.topological.R
@@ -198,6 +198,7 @@ net_par <- function(go, mode = c("v", "e", "n", "all"), fast = TRUE, remove_nega
     # Calculate Vertices Parameters
     v_index <- data.frame(
       check.names = F,
+      name = igraph::vertex_attr(go, "name"),
       Degree = igraph::degree(go),
       `Clustering_coefficient` = igraph::transitivity(go, type = "local"), # local clustering coefficient
       Betweenness = ifelse(negative_weight, igraph::betweenness(up), igraph::betweenness(go)), # betweenness